error printer and some usages
