+++ rust/Cargo.toml
@@ -21,5 +21,5 @@ members = [
     "xet_error",
     "xet_config",
     "xetblob",
-    "lazy",
+    "lazy", "error_printer",
 ]
