+++ rust/error_printer/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "error_printer"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tracing = "0.1.40"
